consolidate some temp types etc
