+++ yarn-project/p2p/src/p2p_client.test.ts
@@ -2,10 +2,13 @@ import { expect, jest } from '@jest/globals';
 
 import { InMemoryP2PCLient } from './memory_p2p_client.js';
 import { TxPool } from './tx_pool.js';
-import { RollupSource } from './types.js';
+import { RollupSource } from './temp_types.js';
 import { MockRollupSource } from './mocks.js';
-import { MockTx } from './tx.js';
+import { MockTx } from './mocks.js';
 
+/**
+ * Mockify helper for testing purposes.
+ */
 type Mockify<T> = {
   [P in keyof T]: ReturnType<typeof jest.fn>;
 };
@@ -25,8 +28,23 @@ describe('In-Memory P2P Client', () => {
     rollupSource = new MockRollupSource();
   });
 
+  it('can start & stop', () => {
+    const client = new InMemoryP2PCLient(rollupSource, txPool);
+    expect(client.isReady()).toEqual(false);
+    expect(client.isRunning()).toEqual(false);
+
+    client.start();
+    expect(client.isReady()).toEqual(true);
+    expect(client.isRunning()).toEqual(true);
+
+    client.stop();
+    expect(client.isReady()).toEqual(false);
+    expect(client.isRunning()).toEqual(false);
+  });
+
   it('adds txs to pool', () => {
     const client = new InMemoryP2PCLient(rollupSource, txPool);
+    client.start();
     const tx1 = new MockTx();
     const tx2 = new MockTx();
     client.sendTx(tx1);
