+++ yarn-project/p2p/src/tx_pool.ts
@@ -1,4 +1,4 @@
-import { Tx } from './tx.js';
+import { Tx } from './temp_types.js';
 
 /**
  * Interface of a transaction pool. The pool includes tx requests and is kept up-to-date by a P2P client.
